+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 - Renamed the app to Exilence (former ExileParty)
 - Improved performance for players with large area-history
 - Improved ladder-functionality
+- Now stores net worth history for two weeks until storage issues have been resolved
 - Fixed rounding of DPS-numbers
 - Fixed a bug where some areas would be missing when parsing the log
 - Changed texts on some buttons in the login-process